refactor: . Move totalTasksCountDisplayText() to new file TaskCountDisplayText.ts
